Document and organise xwrap.rs (Questions) (#511)
* Comment all functions, and standardize comment format

* Move getters

* Order xwrap

* Split xwrap.rs up into files

* Organise files more

* Reduce setters

* Reduce getters

* Add links to xlib calls
